requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
ollama